2009-03-05  Mikael Djurfeldt  <mikael@djurfeldt.com>
	* contdelay.cc (getargs): Use atof instead of atoi for option -d.


git-svn-id: http://svn.incf.org/svn/music/trunk@460 b86cec79-47c1-4bcb-83d2-4e3c60c58a18

+++ test/contdelay.cc
@@ -92,7 +92,7 @@ getargs (int rank, int argc, char* argv[])
 	  timestep = atof (optarg);
 	  continue;
 	case 'd':
-	  delay = atoi(optarg);
+	  delay = atof(optarg);
 	  continue;
 	case 'L':
 	  label = atoi(optarg);
